kernel - Fix vmtotal sysctl
* Recognize unbounded VM objects and do not try to include their sizes
  in the vmtotal structure.

* Should fix systat -vm output for the All/Tot entry.

* Redo the vmstat structure.  Use long's and int64_t's as
  appropriate.

* Adjust systat and vmstat to deal with the new field widths
  in struct vmstat.

Reported-by: Antonio Huete Jimenez <ahuete.devel@gmail.com>

+++ sys/sys/vmmeter.h
@@ -135,20 +135,20 @@ extern struct vmstats vmstats;
 /* systemwide totals computed every five seconds */
 struct vmtotal
 {
-	int16_t	t_rq;		/* length of the run queue */
-	int16_t	t_dw;		/* jobs in ``disk wait'' (neg priority) */
-	int16_t	t_pw;		/* jobs in page wait */
-	int16_t	t_sl;		/* jobs sleeping in core */
-	int16_t	t_sw;		/* swapped out runnable/short block jobs */
-	int32_t	t_vm;		/* total virtual memory */
-	int32_t	t_avm;		/* active virtual memory */
-	int32_t	t_rm;		/* total real memory in use */
-	int32_t	t_arm;		/* active real memory */
-	int32_t	t_vmshr;	/* shared virtual memory */
-	int32_t	t_avmshr;	/* active shared virtual memory */
-	int32_t	t_rmshr;	/* shared real memory */
-	int32_t	t_armshr;	/* active shared real memory */
-	int32_t	t_free;		/* free memory pages */
+	long	t_rq;		/* length of the run queue */
+	long	t_dw;		/* jobs in ``disk wait'' (neg priority) */
+	long	t_pw;		/* jobs in page wait */
+	long	t_sl;		/* jobs sleeping in core */
+	long	t_sw;		/* swapped out runnable/short block jobs */
+	int64_t	t_vm;		/* total virtual memory */
+	int64_t	t_avm;		/* active virtual memory */
+	long	t_rm;		/* total real memory in use */
+	long	t_arm;		/* active real memory */
+	int64_t	t_vmshr;	/* shared virtual memory */
+	int64_t	t_avmshr;	/* active shared virtual memory */
+	long	t_rmshr;	/* shared real memory */
+	long	t_armshr;	/* active shared real memory */
+	long	t_free;		/* free memory pages */
 };
 
 #ifdef PGINPROF
